feat(map config): Improvement
- Improved thematic layer config UI
 - Implemented thematic layer add button Ui and logic
