+++ src/modules/InterventionConfiguration/components/General/components/MapConfiguration/components/ThematicLayer/styles/ThematicConfig.module.css
@@ -1,0 +1,7 @@
+.h-100 {
+    height: 100% !important;
+}
+
+.h-100> div{
+    height: 100% !important;
+}
