+++ src/modules/InterventionConfiguration/components/General/components/MapConfiguration/components/ThematicLayer/styles/thematic-config.css
@@ -7,7 +7,8 @@
     height: 100%;
     display: flex;
     flex-direction: column;
-    gap: 4px;
+    justify-content: space-between;
+    gap: 2px;
 }
 
 .thematic-config-card-header {
